pyproject.toml
[project]
name = "wn-eeg-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-whiteneuron",
    "plotly>=6.3.0",
    "psycopg2>=2.9.10",
    "pycountry>=24.6.1",
    "requests>=2.32.5",]

[tool.uv.sources]
django-whiteneuron = { git = "https://github.com/tanhtm/django-whiteneuron.git", rev = "v0.1.80" }
